backend/requirements.txt
fastapi
requests
sqlalchemy
uvicorn
PyPDF2
python-multipart
reportlab
pytesseract
Pillow
